Move safe PyG loader under loaddata and add MNIST-M load test
Address maintainer review on #580:

- Move kale/utils/serialization.py to kale/loaddata/pyg_serialization.py, since its
  only callers are the loaddata datasets and it is PyG-specific. Update the imports in
  multiomics_datasets and polypharmacy_datasets, the docs entry (kale.utils.rst ->
  kale.loaddata.rst) and the test location.
- Add tests/loaddata/test_mnistm.py: a network-free regression test that builds the
  processed .pt exactly as MNISTM.download() does and confirms it loads under
  weights_only=True and yields a usable sample. This covers the mnistm change outside
  the digits_dann example notebook.

The PyG happy-path is also exercised by the existing MultiomicsDataset/PolypharmacyDataset
tests; the security-specific tests (arbitrary-object rejection and allow-list purity) are
kept here as the regression guard for #541.
